Заготовка 2 таблицы, в базе храниться теперь дробное число, добавлены классы для css

+++ db/migrate/20130103172831_create_finances.rb
@@ -1,7 +1,7 @@
 class CreateFinances < ActiveRecord::Migration
   def change
     create_table :finances do |t|
-      t.integer :m
+      t.decimal :m
       t.text :desc
 
       t.timestamps
